% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemon.R
\name{ipfs_daemon}
\alias{ipfs_daemon}
\title{IPFS daemon}
\usage{
ipfs_daemon(restart = FALSE)
}
\arguments{
\item{restart}{force a restart if ipfs is already running}
}
\description{
Start and stop the ipfs server. This is automatically done when attaching
the package.
}
