% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipns.R
\name{ipfs_resolve}
\alias{ipfs_resolve}
\alias{ipfs_publish}
\title{IPNS name}
\usage{
ipfs_resolve(name = NULL, recursive = FALSE, nocache = FALSE)

ipfs_publish(path, lifetime = "24h")
}
\arguments{
\item{name}{the IPNS name (owner) to resolve. Defaults to your own name}

\item{recursive}{resolve until the result is not an IPNS name}

\item{nocache}{do not use cached entries}

\item{path}{IPFS path of the object to be published under your name}

\item{lifetime}{Time duration that the record will be valid for. This accepts
durations such as "300s", "1.5h" or "2h45m".}
}
\description{
IPNS is a PKI namespace, where names are the hashes of public keys, and
the private key enables publishing new (signed) values. In publish, the
default value of <name> is your own identity public key.
}
\examples{
# Default resolves your own identity:
ipfs_resolve()
ipfs_resolve("QmaCpDMGvV2BGHeYERUEnRQAwe3N8SzbUtfsmvsqQLuvuJ")
ipfs_resolve('ipfs.io')
}
\references{
\url{https://github.com/ipfs/examples/tree/master/examples/ipns#readme}
}

