% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemon.R
\name{ipfs_start}
\alias{ipfs_start}
\alias{ipfs_stop}
\title{IPFS daemon}
\usage{
ipfs_start()

ipfs_stop()
}
\description{
Start and stop the ipfs server. This is automatically done when attaching
the package.
}

