% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfs.R
\name{ipfs}
\alias{ipfs}
\alias{ipfs_get}
\alias{ipfs_version}
\alias{ipfs_webui}
\title{IPFS}
\usage{
ipfs_get(key)

ipfs_version()

ipfs_webui()
}
\arguments{
\item{key}{a multihash key}
}
\description{
Connect to the inter-planetary file system.
}
\examples{
# From 'getting started'
}
\references{
\url{https://ipfs.io/ipfs/QmR7GSQM93Cx5eAg6a6yRzNde1FQv7uL6X1o4k7zrJa3LX/ipfs.draft3.pdf}
ipfs_version()
ipfs_get('QmYwAPJzv5CZsnA625s3Xf2nemtYgPpHdWEz79ojWnPbdG')
}

